Fix string attributes broken by new cgo rules
Strings are unique as they are the only datatype whose concrete type is a
pointer (`char*`). This becomes a problem in Go 1.6 when trying to feed
`H5Awrite` a pointer to the backing array of the original string as it becomes
a pointer to a Go pointer and this [provokes a panic][1].

This commit fixes the problem at the expense of copying the strings. Hopefully
this is not as big a problem for attributes since they tend to be small
compared to datasets.

[1]: https://golang.org/doc/go1.6#cgo