hdf5{,/cmem}: new C data encoder for PT Append()
With this patch, every cgo call, except for the actual call to the HDF5
library was removed. This will improve the performance of this HDF5
wrapper library, as well as leading to more stability in it's execution.
Furthermore, it is not necessary any longer to keep track of the
allocated memory (for both strings and the buffer that has to be
written to the HDF5 PacketTable) and to later free it.

Co-authored-by: Sebastien Binet <binet@cern.ch>